moved edit mode button to drawer menu
