+++ front-src/client/components/Dashboard/Drawer.svelte
@@ -6,6 +6,7 @@
   import clickoutside from "@/libs/svelte/click-outside";
   import HelpLinks from "@/components/App/HelpLinks.svelte";
   import OpenOnStartup from "@/components/App/OpenOnStartup.svelte";
+  import OpenEditMode from "@/components/Panels/OpenEditMode.svelte";
   import LanguageSelect from "@/components/App/LanguageSelect.svelte";
   import OBSConnectAtStartup from "@/components/OBS/ConnectAtStartup.svelte";
   import CheckUpdateCheckbox from "@/components/App/CheckUpdateCheckbox.svelte";
@@ -21,6 +22,7 @@
     style="top:{top}px; max-width:300px;"
     class="absolute z-50 right-0 bottom-0 flex flex-col bg-dark shadow overflow-y-auto"
   >
+    <OpenEditMode on:click="{hide}" />
     <LanguageSelect />
     <div class="flex flex-col divide-y divide-gray-800 divide-opacity-50">
       <OpenOnStartup />
